Move Runtime (type) to runtime (module)
